+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Added the Duffing oscillator in the predefined systems.
 * Changed :bug: where `get_solver` would remove the entry.
 * Added the possibility to provide a state in *all* evolving functions.
+* Added Shinriki oscillator (chaotic 3D flow with period doubling).
 
 # v0.1.0
 Initial release, see: https://juliadynamics.github.io/DynamicalSystems.jl/v0.8.0/